Add support for skipping blank lines
If a CSV file begins with a blank line (before the header), the EasyCSV library will bomb. This is because fgetcsv() returns an array with a single item (with a value of null) when it encounters a blank line. So, in a case where the 1st line is blank, the header row get set to a one-item array. When the 1st real line of data is read (the header row), the array size of the row will not match the single-item array size of $this->_headers, and we get an error.